[Decomp][Spell] Cohesion split: extract crowd-control/immunity aura handlers to SpellAuraControl.cpp
Move the contiguous crowd-control, charm/possess and immunity aura block
(HandleModPossess through HandleAuraModDispelImmunity, 39 handlers) out of
SpellAuras.cpp into a new SpellAuraControl.cpp translation unit.

The file-static frozenAuraTypes[] array is relocated into this TU as well:
its only two consumers (HandleAuraModStun, HandleAuraModRoot) both live in
this block, so it moves with them and is deleted from SpellAuras.cpp.

Pure file-level cohesion split: declarations stay in SpellAuras.h, the
AuraHandler[] dispatch table stays in SpellAuras.cpp and resolves the moved
member methods by address at link time. Byte-identical move, no behaviour
change. SpellAuras.cpp 12994 -> 11576.
